backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
email-validator==2.2.0
alembic==1.16.4
python-dotenv==1.1.1
APScheduler==3.10.4
